lmm/requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pydantic>=2.0.0
requests>=2.28.0
typing-extensions>=4.5.0

# Neural network and machine learning
torch>=2.0.0
transformers>=4.30.0

# Vector database and embeddings
faiss-cpu>=1.7.4  # Use faiss-gpu for CUDA support
chromadb>=0.4.0

# NLP tools
nltk>=3.8.0
textblob>=0.17.0

# Visualization and dashboard
dash>=2.10.0
plotly>=5.14.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0 